fire_and_forget/pyproject.toml
[project]
name = "fire_and_forget"
version = "0.1.0"
description = "fire_and_forget request flow Async RPC example app."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "resonate-sdk==0.4.6",
]

[project.scripts]
foo = "fire_and_forget.service_foo:main"
bar = "fire_and_forget.service_bar:main"
baz = "fire_and_forget.service_baz:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.14.1",
]
